Syncl almost done

+++ Project2/luSYCL.hpp
@@ -1,3 +1,4 @@
 #pragma once
 
-void luBlockFactorizationParallelSYCL(float *a, int n, int blockSize);
+#include <stdlib.h>
+void luBlockFactorizationParallelSYCL(float *a, size_t n, size_t blockSize);
